AI following the player
Lacks Attack mode and animations
To do.
