+++ ggj2019/Assets/Models/room18.fbx.meta
@@ -214,6 +214,44 @@ ModelImporter:
     4300070: Plane.001
     4300072: Cylinder
     4300074: Plane
+    6400000: Cube
+    6400002: Cube.000
+    6400004: Cube.001
+    6400006: Cube.002
+    6400008: Cube.003
+    6400010: Cube.004
+    6400012: Cube.005
+    6400014: Cube.006
+    6400016: Cube.007
+    6400018: Cube.008
+    6400020: Cube.009
+    6400022: Cube.010
+    6400024: Cube.011
+    6400026: Cube.012
+    6400028: Cube.013
+    6400030: Cube.014
+    6400032: Cube.015
+    6400034: Cube.016
+    6400036: Cube.017
+    6400038: Cube.018
+    6400040: Cube.019
+    6400042: Cube.020
+    6400044: Cube.021
+    6400046: Cube.022
+    6400048: Cube.023
+    6400050: Cube.024
+    6400052: Cube.025
+    6400054: Cube.026
+    6400056: Cube.027
+    6400058: Cylinder
+    6400060: Cylinder.001
+    6400062: Cylinder.002
+    6400064: Plane
+    6400066: Plane.001
+    6400068: Plane.002
+    6400070: Plane.003
+    6400072: tracks.007
+    6400074: wase
     10800000: Point
   externalObjects: {}
   materials:
@@ -248,7 +286,7 @@ ModelImporter:
     lODScreenPercentages: []
     globalScale: 1
     meshCompression: 0
-    addColliders: 0
+    addColliders: 1
     useSRGBMaterialColor: 1
     importVisibility: 1
     importBlendShapes: 1
