+++ ggj2019/Assets/Prefabs/doll_enemyv1.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 59424fe69ddb2ad44841aa7cff1d1df1
-folderAsset: yes
-DefaultImporter:
+guid: 9caa71fce98371e408ab2e0b592be6a5
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
