+++ ggj2019/Assets/Scenes/SampleScene1.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 55aee26bc02e1cc4c8f139873e364fcd
+guid: d1f10904e6fc7cd438b12d767613d4eb
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
