+++ ggj2019/Assets/Scenes/SampleScene1/NavMesh.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: eebe446958091a9488f694852d2cba72
-folderAsset: yes
-DefaultImporter:
+guid: 29b34c888c0ced343a4421bcfa61435a
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 23800000
   userData: 
   assetBundleName: 
   assetBundleVariant:
